inline all the things
