--- lib/nextTick.js
@@ -1,11 +1,0 @@
-'use strict';
-exports.test = function () {
-  // Don't get fooled by e.g. browserify environments.
-  return process && !process.browser;
-};
-
-exports.install = function (func) {
-  return function () {
-    process.nextTick(func);
-  };
-};
